Use a dedicated PO parsing and compiling library
The custom methods to generate PO files was not tested enough and were
actually not confronting to the standards. Moreover the library used to
read PO files when used joining with existing translations was old,
unmaintained and faulty too so replace those parts with
`gettext-parser`, a dedicated PO and MO parsing and compiling library.