Added `po-headers` option that can be used to override the default headers.
Example: 

    options["po-headers"] = {
      "project-id-version": "MyProject 1.0",
      "language-team": "MyTeam <team@my.org>",
      "language": "en_US"
    }

Also added a test for custom headers.

Problem to solve:
When jsxgettext is used in an automatic workflow for extracting strings (for instance with grunt-jsxgettext) without the `join-existing` option, the generated .pot file will always contain the default headers.

Since there can be a workflow where the freshly generated .pot file is later merged with .po files, we might want to set project, version and team information already in the .pot file to prevent other tools like poEdit from complaining about missing information.

That's where the `po-headers` option can be used (for instance in the grunt-jsxgettext configuration file).