fix(deps): Use pug instead of jade
The jade module was discontinued and had some `npm audit` warnings.

This replaces jade with pug and updates the jade parser based
on the updated AST format.

There was a problem with the original tests/outputs/jade.pot,
the two modified lines were actually incorrect when comparing
with the source.